+++ crates/ludtwig-parser/CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT-VERSION
 - Fix HTML Tag name parser token collisions, which caused tags like `source` to not parse correctly
 - Fix Twig accessor parser to allow array indexing by dot operator
+- Fix lexer and parsing of colon word syntax like `:special-attribute` as HTML attribute and now support `array[:upper]` as Twig slice
 
 # v0.4.0
 - Rewrite of the whole parser to make it lossless and implemented error recovery